Lua-driven weapon attacks now functioning. Major file organization changes. Cleaned up Entity.java a little bit and combined it with Inventory.java.
