+++ scripts/events/dodge.lua
@@ -1,4 +1,10 @@
+function define()
+	event:addTag("Dodge")
+end
+
+
+
 function invokeEvent()
-	effect = luajava.newInstance("dnd.effects.Effect", "scripts/effects/dodge.lua", "Dodge", source, target)
+	effect = luajava.newInstance("core.Effect", "scripts/effects/dodge.lua", "Dodge", source, target)
 	target:addEffect(effect)
 end
