+++ scripts/notes.txt
@@ -1,0 +1,5 @@
+
+see more: http://www.luaj.org/luaj/3.0/README.html#luajava
+
+vb = luajava.bindClass("dnd.VirtualBoard")
+effect = luajava.newInstance("core.Effect", "scripts/effects/dodge.lua", "Dodge", source, target)
