+++ scripts/tasks/dodge.lua
@@ -1,7 +1,7 @@
 function invokeTask()
-	Event = luajava.bindClass("dnd.events.Event")
-	group = luajava.newInstance("dnd.events.eventgroups.EventGroup")
-	event = luajava.newInstance("dnd.events.Event", "scripts/events/dodge.lua", "Dodge", Event.SINGLE_TARGET)
+	Event = luajava.bindClass("core.events.Event")
+	group = luajava.newInstance("core.events.EventGroup")
+	event = luajava.newInstance("core.events.Event", "scripts/events/dodge.lua", "Dodge", Event.SINGLE_TARGET)
 	group:addEvent(event)
 	invoker:queueEventGroup(group)
 end
