+++ src/core/Observer.java
@@ -1,0 +1,7 @@
+package core;
+
+public interface Observer {
+	
+	public void update(Subject s);
+	
+}
