+++ src/core/Subject.java
@@ -1,0 +1,9 @@
+package core;
+
+public interface Subject {
+	
+	public void addObserver(Observer o);
+	public void removeObserver(Observer o);
+	public void updateObservers();
+	
+}
