--- src/dnd/data/ArmorProperty.java
@@ -1,10 +1,0 @@
-package dnd.data;
-
-public enum ArmorProperty {
-	LIGHT(),
-	MEDIUM(),
-	HEAVY(),
-	STR_13(),
-	STR_15(),
-	STEALTH_DADV();
-}
