--- src/dnd/data/WeaponProperty.java
@@ -1,18 +1,0 @@
-package dnd.data;
-
-public enum WeaponProperty {
-	// Weapon properties
-	AMMUNITION(),
-	FINESSE(),
-	HEAVY(),
-	IMPROVISED_MELEE(),
-	IMPROVISED_THROWN(),
-	LIGHT(),
-	LOADING(),
-	PACT_WEAPON(),
-	RANGE(),
-	REACH(),
-	THROWN(),
-	TWO_HANDED(),
-	VERSATILE();
-}
