--- src/dnd/data/WeaponType.java
@@ -1,47 +1,0 @@
-package dnd.data;
-
-public enum WeaponType {
-	// Weapon categories
-	SIMPLE_MELEE(),
-	SIMPLE_RANGED(),
-	MARTIAL_MELEE(),
-	MARTIAL_RANGED(),
-	// Particular Weapons
-	CLUB(),
-	DAGGER(),
-	GREATCLUB(),
-	HANDAXE(),
-	JAVELIN(),
-	LIGHT_HAMMER(),
-	MACE(),
-	QUARTERSTAFF(),
-	SICKLE(),
-	SPEAR(),
-	LIGHT_CROSSBOW(),
-	DART(),
-	SHORTBOW(),
-	SLING(),
-	BATTLEAXE(),
-	FLAIL(),
-	GLAIVE(),
-	GREATAXE(),
-	GREATSWORD(),
-	HALBERD(),
-	LANCE(),
-	LONGSWORD(),
-	MAUL(),
-	MORNINGSTAR(),
-	PIKE(),
-	RAPIER(),
-	SCIMITAR(),
-	SHORTSWORD(),
-	TRIDENT(),
-	WAR_PICK(),
-	WARHAMMER(),
-	WHIP(),
-	BLOWGUN(),
-	HAND_CROSSBOW(),
-	HEAVY_CROSSBOW(),
-	LONGBOW(),
-	NET();
-}
