--- src/engine/patterns/Observer.java
@@ -1,7 +1,0 @@
-package engine.patterns;
-
-public interface Observer {
-	
-	public void update(Subject s);
-	
-}
