--- src/engine/patterns/Subject.java
@@ -1,9 +1,0 @@
-package engine.patterns;
-
-public interface Subject {
-	
-	public void addObserver(Observer o);
-	public void removeObserver(Observer o);
-	public void updateObservers();
-	
-}
