+++ src/main/Main.java
@@ -1,7 +1,9 @@
 package main;
 
-import gameobjects.entities.Entity;
-import gameobjects.entities.Zombie;
+import core.Item;
+import core.events.groups.EventGroup;
+import core.gameobjects.Entity;
+import core.gameobjects.Zombie;
 import maths.Vector;
 
 public class Main {
@@ -13,9 +15,22 @@ public class Main {
 	private static void test() {
 		Entity z = new Zombie(new Vector(30, 0, 0), new Vector(1, 0, 0));
 		z.invokeTask(0);
-		z.invokeTask(0);
+		for (EventGroup group : z.getEventQueue()) {
+			System.out.println(group);
+		}
 		z.invokeQueuedEvent(0, 0, z.getPos());
+		for (EventGroup group : z.getEventQueue()) {
+			System.out.println(group);
+		}
+		z.equipMainhand(new Item("scripts/items/dagger.lua"));
+		for (EventGroup group : z.getEventQueue()) {
+			System.out.println(group);
+		}
 		z.invokeQueuedEvent(0, 0, z.getPos());
+		z.equipMainhand(new Item("scripts/items/dagger.lua"));
+		for (EventGroup group : z.getEventQueue()) {
+			System.out.println(group);
+		}
 	}
 
 }
