+++ src/testing/dnd/effects/DummyEffect.java
@@ -1,8 +1,8 @@
 package testing.dnd.effects;
 
-import dnd.effects.Effect;
-import dnd.events.Event;
-import gameobjects.entities.Entity;
+import core.Effect;
+import core.events.Event;
+import core.gameobjects.Entity;
 
 public class DummyEffect extends Effect {
