+++ src/testing/gameobjects/entities/DummyEntity.java
@@ -1,6 +1,6 @@
 package testing.gameobjects.entities;
 
-import gameobjects.entities.Entity;
+import core.gameobjects.Entity;
 import maths.Vector;
 
 public class DummyEntity extends Entity {
